Refactor repo into src package layout
