+++ src/nxtgenai/__init__.py
@@ -1,0 +1,1 @@
+"""NxtGenAI package."""
